+++ threads/dft-vrank-geq1.c
@@ -18,7 +18,7 @@
  *
  */
 
-/* $Id: dft-vrank-geq1.c,v 1.22 2006-01-05 23:23:15 athena Exp $ */
+/* $Id: dft-vrank-geq1.c,v 1.23 2006-01-10 13:24:41 athena Exp $ */
 
 #include "threads.h"
 
@@ -221,7 +221,7 @@ void X(dft_thr_vrank_geq1_register)(planner *p)
      /* FIXME: Should we try other vecloop_dim values? */
      static const int buddies[] = { 1, -1 };
 
-     const int nbuddies = sizeof(buddies) / sizeof(buddies[0]);
+     const int nbuddies = (int)(sizeof(buddies) / sizeof(buddies[0]));
 
      for (i = 0; i < nbuddies; ++i)
           REGISTER_SOLVER(p, mksolver(buddies[i], buddies, nbuddies));
